Only a small subset of KML objects may actually be used in a Delete, so I made a Deletable interface and moved the writeDelete method into it. The Document, Folder, Placemark, GroundOverlay, and ScreenOverlay types were update to implement Deletable and all other types were modified to remove the writeDelete method. The Delete class was modified to call the method in Deletable and to only accept Deletable values for its kmlObject field.
Also corrected several "visibility" misspellings.  Added support in the Kml class for the extension schema and added a printNoIndent method to make coordinates easier to read in output.

In Model, I added the ability to specify an ID for the location and orientation so that these can be modified with Update Change later.